+++ code/crates/test/tests/it/value_sync.rs
@@ -200,6 +200,42 @@ pub async fn start_late() {
 }
 
 #[tokio::test]
+pub async fn start_late_parallel_requests() {
+    const HEIGHT: u64 = 10;
+
+    let mut test = TestBuilder::<()>::new();
+
+    test.add_node()
+        .with_voting_power(10)
+        .start()
+        .wait_until(HEIGHT * 2)
+        .success();
+
+    test.add_node()
+        .with_voting_power(10)
+        .start()
+        .wait_until(HEIGHT * 2)
+        .success();
+
+    test.add_node()
+        .with_voting_power(5)
+        .start_after(1, Duration::from_secs(10))
+        .wait_until(HEIGHT)
+        .success();
+
+    test.build()
+        .run_with_params(
+            Duration::from_secs(30),
+            TestParams {
+                enable_value_sync: true,
+                parallel_requests: 5,
+                ..Default::default()
+            },
+        )
+        .await
+}
+
+#[tokio::test]
 pub async fn start_late_rotate_epoch_validator_set() {
     const HEIGHT: u64 = 20;